update crud request

+++ Prueba/controllers/componentController.js
@@ -5,35 +5,30 @@ const fs = require('fs');
 
 // Crear un componente
 const createComponent = async (req, res) => {
-  upload.single('image')(req, res, async (err) => {
-    if (err) {
-      return res.status(400).json({ error: err.message });
+  try {
+    const data = req.body;
+    
+    if (req.file) {
+      data.imageUrl = `/uploads/${req.file.filename}`;
     }
-    try {
-      const data = req.body;
-      if (req.file) {
-        data.imageUrl = `/uploads/${req.file.filename}`;
-      }
-      const component = await componentModel.createComponent(data);
-      res.status(201).json(component);
-    } catch (error) {
-      console.error('Error al crear el componente:', error.message);
-      res.status(500).json({ error: 'Error al crear el componente' });
-    }
-  });
+    const component = await componentModel.createComponent(data);
+    res.status(201).json(component);
+  } catch (error) {
+    console.error('Error al crear el componente:', error);
+    res.status(500).json({ error: error.message });
+  }
 };
 
-
-// Obtener todos los componentes
+// Obtener todos los componentes con filtro por estado
 const getAllComponents = async (req, res) => {
-  const { name } = req.query;
+  const { name, status } = req.query;
 
   try {
     let components;
     if (name) {
       components = await componentModel.searchComponentsByName(name);
     } else {
-      components = await componentModel.getAllComponents();
+      components = await componentModel.getAllComponents(status);  // Pasamos el filtro de estado
     }
 
     res.status(200).json({ components });
@@ -41,6 +36,7 @@ const getAllComponents = async (req, res) => {
     res.status(500).json({ error: error.message });
   }
 };
+
 
 // Obtener un componente por ID
 const getComponentById = async (req, res) => {
@@ -58,34 +54,29 @@ const getComponentById = async (req, res) => {
 
 // Actualizar un componente por ID
 const updateComponent = async (req, res) => {
-  upload.single('image')(req, res, async (err) => {
-    if (err) {
-      return res.status(400).json({ error: err.message });
-    }
-    const { id } = req.params;
-    const data = req.body;
-    try {
-      const currentComponent = await componentModel.getComponentById(id);
-      if (!currentComponent) {
-        return res.status(404).json({ error: 'Componente no encontrado' });
-      }
-      if (req.file) {
-        if (currentComponent.imageUrl) {
-          const oldImagePath = path.join(__dirname, '..', 'uploads', path.basename(currentComponent.imageUrl));
-          if (fs.existsSync(oldImagePath)) {
-            fs.unlinkSync(oldImagePath);
-          }
-        }
-        data.imageUrl = `/uploads/${req.file.filename}`;
-      }
+  const { id } = req.params;
+  const data = req.body;
 
-      const updatedComponent = await componentModel.updateComponent(id, data);
-      res.status(200).json(updatedComponent);
-    } catch (error) {
-      console.error('Error al actualizar el componente:', error.message);
-      res.status(500).json({ error: 'Error al actualizar el componente' });
+  try {
+    const currentComponent = await componentModel.getComponentById(id);
+    if (!currentComponent) {
+      return res.status(404).json({ error: 'Componente no encontrado' });
     }
-  });
+    if (req.file) {
+      const oldImagePath = path.join(__dirname, '..', 'uploads', path.basename(currentComponent.imageUrl));
+      data.imageUrl = `/uploads/${req.file.filename}`;
+      if (currentComponent.imageUrl && fs.existsSync(oldImagePath)) {
+        fs.unlinkSync(oldImagePath);
+      }
+    } else {
+      data.imageUrl = currentComponent.imageUrl;
+    }
+    const updatedComponent = await componentModel.updateComponent(id, data);
+    res.status(200).json(updatedComponent);
+  } catch (error) {
+    console.error('Error al actualizar el componente:', error.message);
+    res.status(500).json({ error: error.message });
+  }
 };
 
 // Eliminar un componente por ID
@@ -116,7 +107,7 @@ const filterComponentsByCategories = async (req, res) => {
   const { categoryIds } = req.query; 
   try {
     if (!categoryIds) {
-      return res.status(400).json({ error: 'Se deben proporcionar IDs de categorías' });
+      return res.status(400).json({ error: 'Se deben proporcionar IDs de categorÃ­as' });
     }
     const components = await componentModel.filterComponentsByCategories(categoryIds);
     res.status(200).json(components);
@@ -135,6 +126,7 @@ const getComponentCount = async (req, res) => {
     res.status(500).json({ error: error.message });
   }
 };
+
 
 module.exports = {
   createComponent,
